# Support optimized witness generation for circuit2.