--- a/benches/synthesis.rs
+++ b/benches/synthesis.rs
@@ -1,44 +1,56 @@
 use crate::poseidon::{Arity, PoseidonConstants};
 use bellperson::gadgets::num::AllocatedNum;
 use bellperson::util_cs::bench_cs::BenchCS;
-use bellperson::{Circuit, ConstraintSystem, SynthesisError};
+use bellperson::util_cs::witness_cs::WitnessCS;
+use bellperson::{ConstraintSystem, SynthesisError};
 use blstrs::Scalar as Fr;
 use criterion::{criterion_group, criterion_main, BenchmarkId, Criterion};
 use ff::Field;
 use generic_array::typenum;
 use neptune::circuit::{poseidon_hash_circuit, CircuitType};
+use neptune::circuit2::poseidon_hash_allocated_mut;
 use neptune::*;
 use rand::thread_rng;
 use std::marker::PhantomData;
 
 struct BenchCircuit<'a, A: Arity<Fr>> {
     n: usize,
-    circuit_type: &'a CircuitType,
+    circuit_type: Option<&'a CircuitType>,
     _a: PhantomData<A>,
 }
 
-impl<A: Arity<Fr>> Circuit<Fr> for BenchCircuit<'_, A> {
-    fn synthesize<CS: ConstraintSystem<Fr>>(self, mut cs: &mut CS) -> Result<(), SynthesisError> {
-        let mut rng = thread_rng();
-        let arity = A::to_usize();
-        let constants = PoseidonConstants::<Fr, A>::new();
-
+impl<A: Arity<Fr>> BenchCircuit<'_, A> {
+    fn synthesize<CS: ConstraintSystem<Fr>>(
+        self,
+        cs: &mut CS,
+        data: Vec<AllocatedNum<Fr>>,
+        constants: &PoseidonConstants<Fr, A>,
+    ) -> Result<(), SynthesisError> {
         for _ in 0..self.n {
-            let mut i = 0;
-            let mut fr_data = vec![Fr::random(&mut rng); arity];
-            let data: Vec<AllocatedNum<Fr>> = (0..arity)
-                .enumerate()
-                .map(|_| {
-                    let fr = Fr::random(&mut rng);
-                    fr_data[i] = fr;
-                    i += 1;
-                    AllocatedNum::alloc(cs.namespace(|| format!("data {}", i)), || Ok(fr)).unwrap()
-                })
-                .collect::<Vec<_>>();
-            let _ = poseidon_hash_circuit(&mut cs, *self.circuit_type, data, &constants)
-                .expect("poseidon hashing failed");
+            let _ = if self.circuit_type.is_some() {
+                poseidon_hash_circuit(cs, *self.circuit_type.unwrap(), data.clone(), constants)
+                    .expect("poseidon hashing failed");
+            } else {
+                poseidon_hash_allocated_mut(cs, data.clone(), constants)
+                    .expect("poseidon hashing failed");
+            };
         }
         Ok(())
+    }
+
+    fn data<CS: ConstraintSystem<Fr>>(cs: &mut CS) -> Vec<AllocatedNum<Fr>> {
+        let mut rng = thread_rng();
+        let arity = A::to_usize();
+
+        let mut fr_data = vec![Fr::random(&mut rng); arity];
+        let data: Vec<AllocatedNum<Fr>> = (0..arity)
+            .map(|i| {
+                let fr = Fr::random(&mut rng);
+                fr_data[i] = fr;
+                AllocatedNum::alloc(cs.namespace(|| format!("data {}", i)), || Ok(fr)).unwrap()
+            })
+            .collect::<Vec<_>>();
+        data
     }
 }
 
@@ -47,6 +59,7 @@ where
     A: Arity<Fr>,
 {
     let mut group = c.benchmark_group(format!("synthesis-{}", A::to_usize()));
+    let constants = PoseidonConstants::<Fr, A>::new();
     for i in 0..4 {
         let num_hashes = 10usize.pow(i);
         for circuit_type in &[CircuitType::Legacy, CircuitType::OptimalAllocated] {
@@ -57,19 +70,39 @@ where
                 ),
                 &num_hashes,
                 |b, n| {
+                    let mut cs = BenchCS::<Fr>::new();
+                    let data = BenchCircuit::<A>::data(&mut cs);
                     b.iter(|| {
-                        let mut cs = BenchCS::<Fr>::new();
                         let circuit = BenchCircuit::<A> {
                             n: *n,
-                            circuit_type,
+                            circuit_type: Some(circuit_type),
                             _a: PhantomData::<A>,
                         };
-                        circuit.synthesize(&mut cs)
+                        circuit.synthesize(&mut cs, data.clone(), &constants)
                     })
                 },
             );
         }
-        // num_hashes *= 10;
+
+        group.bench_with_input(
+            BenchmarkId::new(
+                "hash_allocated_witness",
+                format!("arity: {}, count: {}", A::to_usize(), num_hashes),
+            ),
+            &num_hashes,
+            |b, n| {
+                let mut cs = WitnessCS::<Fr>::new();
+                let data = BenchCircuit::<A>::data(&mut cs);
+                b.iter(|| {
+                    let circuit = BenchCircuit::<A> {
+                        n: *n,
+                        circuit_type: None,
+                        _a: PhantomData::<A>,
+                    };
+                    circuit.synthesize(&mut cs, data.clone(), &constants)
+                })
+            },
+        );
     }
 }