--- a/benches/synthesis.rs
+++ b/benches/synthesis.rs
@@ -1,8 +1,8 @@
 use crate::poseidon::{Arity, PoseidonConstants};
-use bellperson::gadgets::num::AllocatedNum;
-use bellperson::util_cs::bench_cs::BenchCS;
-use bellperson::util_cs::witness_cs::WitnessCS;
-use bellperson::{ConstraintSystem, SynthesisError};
+use bellpepper::gadgets::num::AllocatedNum;
+use bellpepper::util_cs::bench_cs::BenchCS;
+use bellpepper::util_cs::witness_cs::WitnessCS;
+use bellpepper::{ConstraintSystem, SynthesisError};
 use blstrs::Scalar as Fr;
 use criterion::{criterion_group, criterion_main, BenchmarkId, Criterion};
 use ff::Field;