--- a/src/circuit2_witness.rs
+++ b/src/circuit2_witness.rs
@@ -234,7 +234,10 @@ where
 
                 // sparse mds
                 {
-                    elements_buffer[0] = scalar_product(elements, &m.w_hat);
+                    elements_buffer[0] = elements
+                        .iter()
+                        .zip(&m.w_hat)
+                        .fold(Scalar::ZERO, |acc, (&x, &y)| acc + (x * y));
 
                     for j in 1..width {
                         elements_buffer[j] = elements[j] + elements[0] * m.v_rest[j - 1];
@@ -313,12 +316,6 @@ where
 
         elements[1]
     }
-}
-
-fn scalar_product<Scalar: PrimeField>(a: &[Scalar], b: &[Scalar]) -> Scalar {
-    a.iter()
-        .zip(b)
-        .fold(Scalar::ZERO, |acc, (&x, &y)| acc + (x * y))
 }
 
 #[cfg(test)]