# [Do not merge until updated for bellpepper.]

**UPDATE**: This PR now uses [bellpepper](https://github.com/lurk-lab/bellpepper). However, we intend to retain bellperson compatibilty for those requiring it. This PR switches to bellpepper to expedite upstream changes. Therefore it should also add a mechanism for using bellperson instead, when needed.

**UPDATE2**: I have added a `bellperson` feature, so now this only needs to be updated to exercise the feature flag on CI.

**UPDATE3**: Actually, the new work still depends on `bellperson` git dependency, so maybe the answer is to make the new work unsupported when `bellperson` feature is selected.

This PR makes use of https://github.com/filecoin-project/bellperson/pull/309 to implement `SizedWitness` for `poseidon`. The resulting witness is for the `circuit2` variant, with final result allocated. Several entry points are provided, to support usages where the witness should be, for example, generated directly into the constraint system, or returned for later use.

The synthesis benchmark is extended to allow comparison with previous synthesis (using `BenchCS`, which carries minimal overhead). Note that the `hash_allocated_witness` runs are about 16x faster than the `Optimal Allocated Poseidon Circuit` runs.

```
➜  neptune git:(witness-gen) ✗ cargo criterion --bench synthesis
   Compiling neptune v9.0.0 (/Users/clwk/fil/neptune)
    Finished bench [optimized] target(s) in 5.82s
synthesis-8/Legacy Poseidon Circuit/arity: 8, count: 1                                                                            
                        time:   [606.56 µs 608.75 µs 612.32 µs]
                        change: [-0.6349% -0.2242% +0.2068%] (p = 0.37 > 0.05)
                        No change in performance detected.
synthesis-8/Optimal Allocated Poseidon Circuit,/arity: 8, count: 1                                                                            
                        time:   [591.16 µs 591.63 µs 592.15 µs]
                        change: [-1.3293% -0.5826% -0.1104%] (p = 0.08 > 0.05)
                        No change in performance detected.
synthesis-8/hash_allocated_witness/arity: 8, count: 1                                                                            
                        time:   [36.666 µs 36.695 µs 36.747 µs]
                        change: [+0.6121% +0.8506% +1.1294%] (p = 0.00 < 0.05)
                        Change within noise threshold.
synthesis-8/Legacy Poseidon Circuit/arity: 8, count: 10                                                                           
                        time:   [6.1382 ms 6.1424 ms 6.1468 ms]
                        change: [-1.9551% -1.0453% -0.2484%] (p = 0.02 < 0.05)
                        Change within noise threshold.
synthesis-8/Optimal Allocated Poseidon Circuit,/arity: 8, count: 10                                                                           
                        time:   [5.9172 ms 5.9187 ms 5.9217 ms]
                        change: [+0.8503% +1.1880% +1.5227%] (p = 0.00 < 0.05)
                        Change within noise threshold.
synthesis-8/hash_allocated_witness/arity: 8, count: 10                                                                           
                        time:   [365.80 µs 366.54 µs 367.76 µs]
                        change: [-15.570% -14.962% -14.400%] (p = 0.00 < 0.05)
                        Performance has improved.
synthesis-8/Legacy Poseidon Circuit/arity: 8, count: 100                                                                           
                        time:   [61.153 ms 61.457 ms 61.860 ms]
                        change: [+2.2618% +2.5457% +2.9353%] (p = 0.00 < 0.05)
                        Performance has regressed.
synthesis-8/Optimal Allocated Poseidon Circuit,/arity: 8, count: 100                                                                           
                        time:   [58.389 ms 58.418 ms 58.446 ms]
                        change: [-32.491% -32.425% -32.362%] (p = 0.00 < 0.05)
                        Performance has improved.
synthesis-8/hash_allocated_witness/arity: 8, count: 100                                                                            
                        time:   [3.6248 ms 3.6411 ms 3.6632 ms]
                        change: [-95.978% -95.962% -95.943%] (p = 0.00 < 0.05)
                        Performance has improved.
Benchmarking synthesis-8/Legacy Poseidon Circuit/arity: 8, count: 1000: Warming up for 3.0000 s
Warning: Unable to complete 10 samples in 5.0s. You may wish to increase target time to 6.1s.
synthesis-8/Legacy Poseidon Circuit/arity: 8, count: 1000                                                                          
                        time:   [605.64 ms 607.31 ms 609.05 ms]
                        change: [+1.1024% +1.3995% +1.7238%] (p = 0.00 < 0.05)
                        Performance has regressed.
Benchmarking synthesis-8/Optimal Allocated Poseidon Circuit,/arity: 8, count: 1000: Warming up for 3.0000 s
Warning: Unable to complete 10 samples in 5.0s. You may wish to increase target time to 5.9s.
synthesis-8/Optimal Allocated Poseidon Circuit,/arity: 8, count: 1000                                                                          
                        time:   [585.13 ms 587.99 ms 591.48 ms]
                        change: [-32.349% -31.993% -31.619%] (p = 0.00 < 0.05)
                        Performance has improved.
synthesis-8/hash_allocated_witness/arity: 8, count: 1000                                                                           
                        time:   [36.332 ms 36.387 ms 36.419 ms]
                        change: [-95.984% -95.973% -95.962%] (p = 0.00 < 0.05)
                        Performance has improved.
```