@porcuquine 2023-06-20
I pushed a commit that makes the witness-generation optimization automatically available to SAFE.