@huitseeker 2023-06-22
FYI: https://github.com/lurk-lab/neptune/pull/199 is changing the required tests upon merging (removing “rustfmt”, which is now integrated in “clippy” and adding “msrv”, which requires source compatibility with the declared rust-version). The PR in its current form has not run the second check, hence the “Some checks haven’t completed yet”, but that should resolve when you rebase.